verificando se existe path e adicionando jest

+++ src/helpers/fileHelpers.js
@@ -35,21 +35,23 @@ function compressFiles() {
   }
 }
 
-function removeFiles() {
+async function removeFiles() {
   const directory = getFilePath('../public/files');
   if (fs.existsSync(getFilePath('../public/files.zip'))) {
     fs.unlinkSync(getFilePath('../public/files.zip'));
   }
 
-  fs.readdir(directory, (err, files) => {
-    if (err) throw err;
+  if (fs.existsSync(directory)) {
+    fs.readdir(directory, (err, files) => {
+      if (err) throw err;
 
-    for (const file of files) {
-      fs.unlink(path.join(directory, file), (err) => {
-        if (err) throw err;
-      });
-    }
-  });
+      for (const file of files) {
+        fs.unlink(path.join(directory, file), (err) => {
+          if (err) throw err;
+        });
+      }
+    });
+  }
 }
 
 function getFilePath(currentPath) {
